Initial version of evict-cache plugin
This plugin allows to synchronize the eviction of caches between two
Gerrit instances sharing the same repositories and database.

The plugin needs to be installed in both instances and every time a
cache eviction occurs in one of the instances, the other instance's
cache is updated. This way, both caches are kept synchronized.

Synchronized caches are:
  - accounts
  - accounts_byemail
  - accounts_byname
  - adv_bases
  - groups
  - groups_byinclude
  - groups_byname
  - groups_byuuid
  - groups_external
  - groups_members
  - ldap_groups
  - ldap_usernames
  - project_list
  - projects
  - sshkeys
  - web_sessions

Change-Id: I4f5c1dc5c705854ae6af798f392c681078355a6c